+++ CHANGELOG.md
@@ -21,11 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Space semantic tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Tag component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Link component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Chip component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
-- Button component (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
-- Button component (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- Button component (Orange-OpenSource/ouds-ios#875)
+- Button component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Bullet list component tokens for Orange theme (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Split color charts semantic tokens to dedicated provider (Orange-OpenSource/ouds-ios#856)
 - Swift package `swift-docc-plugin` from v1.4.3 to v1.4.5